requirements.txt
Flask==3.0.2
Flask-MySQLdb==1.0.1
PyMySQL==1.1.0
bcrypt==4.1.2
python-dotenv==1.0.1
gunicorn==21.2.0
cryptography==42.0.5
